pyproject.toml
[tool.poetry]
name = "do"
version = "0.1.0"
description = ""
authors = ["Chris Griffith <chris@cdgriffith.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-box = "^3.2"
"ruamel.yaml" = "^0.15.86"
python-digitalocean = "^1.14"
cutie = "^0.2.2"
appdirs = "^1.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
